+++ src/modules/authorization/utils.ts
@@ -1,4 +1,5 @@
 import { ethers } from 'ethers'
+import { Provider } from '@ethersproject/providers'
 import { Authorization, AuthorizationType } from './types'
 
 export enum AuthorizationError {
@@ -68,4 +69,30 @@ export function areEqual(left: Authorization, right: Authorization) {
 
 export function isValidType(type: string) {
   return Object.values<string>(AuthorizationType).includes(type)
+}
+
+export function getERC20ContractInstance(
+  contractAddress: string,
+  provider: Provider
+) {
+  return new ethers.Contract(
+    contractAddress,
+    [
+      'function allowance(address owner, address spender) view returns (uint256)'
+    ],
+    provider
+  )
+}
+
+export function getERC721ContractInstance(
+  contractAddress: string,
+  provider: Provider
+) {
+  return new ethers.Contract(
+    contractAddress,
+    [
+      'function isApprovedForAll(address owner, address operator) view returns (bool)'
+    ],
+    provider
+  )
 }
